Update article spec
